requirements.txt
requests>=2.32.3,<3.0.0
html2text>=2020.1.16,<2021.0.0
PyYAML>=6.0.1,<7.0.0
